FASE_3/CapTransportation/requirements.txt
Flask==3.0.3
gunicorn==21.2.0

requests==2.32.3

SQLAlchemy==2.0.32
Flask-SQLAlchemy==3.1.1
psycopg2-binary

PyJWT==2.10.1
python-dotenv==1.0.1

# Reports
xhtml2pdf==0.2.17
reportlab==4.4.7
